Conflicts:
	src/System.Text.RegularExpressions/tests/project.json

+++ src/System.Diagnostics.Process/src/System/Diagnostics/ProcessThread.Windows.cs
@@ -171,7 +171,7 @@ namespace System.Diagnostics
         private SafeThreadHandle OpenThreadHandle(int access)
         {
             EnsureState(State.IsLocal);
-            return ProcessManager.OpenThread(_threadInfo._threadId, access);
+            return ProcessManager.OpenThread((int)_threadInfo._threadId, access);
         }
     }
 }
